+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 All notable changes to this project are documented in this file.
 
-## [Version] - Date
+## [1.1.0] - 2024-07-03
 
-### Action
+### Added
+
+- Added `ErrorOr.of(ErrorOr<U> that)` to create a copy of a value instance.
+- Added `ErrorOr.ofError(ErrorOr<U> that)` to create a copy of an error instance.